simplify modailties further
